+++ docs/changelog.rst
@@ -13,6 +13,7 @@ Develop
 * [BUGFIX] Display correct unexpected_percent in DataDocs - corrects the result object from map expectations to return the same "unexpected_percent" as is used to evaluate success (excluding null values from the denominator). The old value is now returned in a key called "unexpected_percent_total" (thanks @mlondschien) #1875
 * [BUGFIX] Add python=3.7 argument to conda env creation (thanks @scouvreur!) #2391
 * [MAINTENANCE] Add checkpoint store to store backend defaults #2378
+* [ENHANCEMENT] Removed the logic that prints the "This configuration object was built using version..." warning when current version of Great Expectations is not the same as the one used to build the suite, since it was not actionable #2366
 
 0.13.8
 -----------------